rest_core/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers

from .models import Trip


class TripSerializer(serializers.ModelSerializer):

    class Meta:
        model = Trip
        fields = ('name', 'start', 'end')


class UserSerializer(serializers.HyperlinkedModelSerializer):

    class Meta:
        model = User
        fields = ('url', 'username', 'email')
